+++ R/deprec-extract-info.R
@@ -9,7 +9,7 @@
 #' `r lifecycle::badge("deprecated")`
 #'
 #' Functionality for accessing recent extract information has been moved to
-#' [`get_recent_extracts_info()`].
+#' [`get_extract_history()`].
 #'
 #' Please use that function instead.
 #'
@@ -22,12 +22,13 @@ get_recent_extracts_info_list <- function(collection = NULL,
   lifecycle::deprecate_warn(
     "0.6.0",
     "get_recent_extracts_info_list()",
-    "get_recent_extracts_info()"
+    "get_extract_history()"
   )
 
-  get_recent_extracts_info(
+  get_extract_history(
     collection = collection,
     how_many = how_many,
+    as_table = FALSE,
     api_key = api_key
   )
 }
@@ -40,13 +41,16 @@ get_recent_extracts_info_tbl <- function(collection = NULL,
   lifecycle::deprecate_warn(
     "0.6.0",
     "get_recent_extracts_info_tbl()",
-    details = "Please use `get_recent_extracts_info()` with `table = TRUE` instead."
+    details = paste0(
+      "Please use `get_extract_history()` with `as_table = TRUE` ",
+      "instead."
+    )
   )
 
-  get_recent_extracts_info(
+  get_extract_history(
     collection = collection,
     how_many = how_many,
-    table = TRUE,
+    as_table = TRUE,
     api_key = api_key
   )
 }